% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSure-package.R
\docType{package}
\name{BSure-package}
\alias{BSure-package}
\alias{BSure}
\title{The 'BSure' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
