% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{compare_length_of_credible_intervals}
\alias{compare_length_of_credible_intervals}
\title{compare_length_of_credible_intervals
Compares the length of the credible intervals of the gene essentiality estimates for two different screens or
studies.}
\usage{
compare_length_of_credible_intervals(
  extracted_output_1,
  extracted_output_2,
  study_name_1,
  study_name_2
)
}
\arguments{
\item{extracted_output_1}{extracted output for screen/study 1, computed from BSure output using the function
extract_from_output.}

\item{extracted_output_2}{extracted output for screen/study 2, computed from BSure output using the function
extract_from_output.}

\item{study_name_1}{name of first screen/study}

\item{study_name_2}{name of second screen/study}
}
\value{
ratio of credible intervals, prints scatterplot comparing lengths of credible intervals.
}
\description{
compare_length_of_credible_intervals
Compares the length of the credible intervals of the gene essentiality estimates for two different screens or
studies.
}
