% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{compare_to_exisiting_screen}
\alias{compare_to_exisiting_screen}
\title{compare_to_exisiting_screen
Compares a new screen to your gold-standard screen. For example, your gold-standard screen may be of high
coverage and your new screen of lower coverage, and you want to check whether essential genes of type I or II
from the gold standard screen are also recovered by the new screen.}
\usage{
compare_to_exisiting_screen(extracted_output_1, extracted_output_2)
}
\value{
proportion_genes_recovered Proportion of genes essential of both type I and II in the first screen that
is essential of either type in the second screen.
}
\description{
compare_to_exisiting_screen
Compares a new screen to your gold-standard screen. For example, your gold-standard screen may be of high
coverage and your new screen of lower coverage, and you want to check whether essential genes of type I or II
from the gold standard screen are also recovered by the new screen.
}
