% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{compare_to_gold_standard}
\alias{compare_to_gold_standard}
\title{compare_to_gold_standard
Compares a new screen to your gold-standard screen. For example, your gold-standard screen may be of high
coverage and your new screen of lower coverage, and you want to check whether essential genes of type I or II
from the gold standard screen are also recovered by the new screen.}
\usage{
compare_to_gold_standard(extracted_output_gold_standard, extracted_output_new)
}
\value{
proportion_genes_recovered_I Proportion of the genes essential of type I in the first screen that are
also essential of type I in the second screen.

proportion_genes_recovered_II Proportion of the genes essential of type II in the first screen that are
also essential of type I in the second screen.

proportion_genes_recovered_I_FNR Proportion of the genes essential of type I in the first screen that are
also essential of type I in the second screen, when FNR is controlled at 0.05 in the second screen. To avoid cases with
very low FNR we use the union of the FNR and FDR controlled sets.

proportion_genes_recovered_II_FNR Proportion of the genes essential of type II in the first screen that are
also essential of type I in the second screen, when FNR is controlled at 0.05 in the second screen. Again we use
the union of the FNR and FDR controlled sets.
}
\description{
compare_to_gold_standard
Compares a new screen to your gold-standard screen. For example, your gold-standard screen may be of high
coverage and your new screen of lower coverage, and you want to check whether essential genes of type I or II
from the gold standard screen are also recovered by the new screen.
}
