% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{compute_lfc_from_counts}
\alias{compute_lfc_from_counts}
\title{compute_lfc_from_counts}
\usage{
compute_lfc_from_counts(counts, controls)
}
\arguments{
\item{counts}{MxR matrix of gRNA counts, where M is the number of gRNAs and R the number of replicates}

\item{controls}{Mx1 vector or MxC matrix, if there are C different control measurements}
}
\value{
matrix of log-fold changes
}
\description{
Computes log-fold changes of gRNA counts from gRNA count and control data.
}
\author{
Magdalena Strauss
}
