% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{compute_mean_lfc}
\alias{compute_mean_lfc}
\title{compute_mean_lfc}
\usage{
compute_mean_lfc(lfc_matrix, gene_vector)
}
\arguments{
\item{lfc_matrix}{MxR matrix of log-fold changes, where M is the number of gRNAs and R the number of replicates}

\item{gene_vector}{vector of length M listing the genes corresponding to the gRNAs (in the same order as the rows of lfc)}
}
\value{
vector of mean log-fold changes across replicates
}
\description{
compute_mean_lfc
}
\author{
Magdalena Strauss
}
