% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{create_small_subset}
\alias{create_small_subset}
\title{create_small_subset
This function takes as input a matrix of count data from a CRISPR screen
and returns a smaller subset containing 20 core essential and 20 nonessential genes,
as well as 20 genes known to be differentially essential across different (cancer) cell lines
(Behan et al., 2019). The purpose of this function is to create a small sample dataset to illustrate
the method.}
\usage{
create_small_subset(counts)
}
\arguments{
\item{counts}{Mx(R+2) matrix or data frame of gRNA counts, where M is the number of gRNAs and R the number of replicates,
the first column of the matrix contains the names of the gRNAs, the second gene names or identifiers
corresponding to the gRNAs in the first column}
}
\value{
reduced matrix of gRNA counts, containing 20 core-essential genes, 20 nonessential genes, and 20
genes know to be differentially essential across different (cancer) cell lines
}
\description{
create_small_subset
This function takes as input a matrix of count data from a CRISPR screen
and returns a smaller subset containing 20 core essential and 20 nonessential genes,
as well as 20 genes known to be differentially essential across different (cancer) cell lines
(Behan et al., 2019). The purpose of this function is to create a small sample dataset to illustrate
the method.
}
\examples{
data("counts_HT29")
counts_HT29_small_subset <- create_small_subset(counts_HT29)
}
\references{
Behan FM et al. Prioritization of cancer therapeutic targets using CRISPR–Cas9 screens.
Nature. 2019;568(7753):511–516.
}
