% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{extract_from_output}
\alias{extract_from_output}
\title{extract_from_output}
\usage{
extract_from_output(BSure_output)
}
\arguments{
\item{BSure_output}{output created by BSure function}
}
\value{
sample_summary with mean estimates (mean) and 95\% credible intervals (CI_lower_bounds
for the lower bounds of the intervals, CI_upper bounds for the upper ones) of the essentiality
level of each gene, the probability of a gene being highly essential (prob_essential_II; the posterior distribution
of the essentiality being shifted less than 1/3 compared to a typical core-essential gene), the
probability of a gene being essential of level I (no overlap with distribution of typical nonessential gene),
and convergence statistics (Rhat=Gelman-Rubin R-hat statistic, tail_ESS=tail essential sample size)
In addition to the sample summary: prop_expensive_sampling: proportion of genes which
required a more expensive sampling algorithm
prop_very_expensive_sampling: proportion of genes which required the most expensive sampling
algorithm
}
\description{
Extracts core output from the output file created by the BSure function.
The extracted output is used in postprocessing functions for plotting and quality control.
}
