% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{extract_gene_categories}
\alias{extract_gene_categories}
\title{extract_gene_categories}
\usage{
extract_gene_categories(
  extracted_output,
  FDR_II = 0.05,
  FDR_I = 0.05,
  figures = T
)
}
\arguments{
\item{extracted_output}{output of the function extract_from_output}

\item{FDR_II}{set to 0.05 by default; false positive rate for essentiality of type II}

\item{FDR_I}{false positive rate for essentiality of type I; set to 0.05 by default}
}
\value{
essential_genes_II  genes of essentiality level II (see above)
essential_genes_I genes of essentiality level I (see above).
proportion_core_essential_in_essential_II which proportion of core essential genes is in the essential genes
II group
(as core essential genes we use the intersection of core essential genes from  Hart and Moffat (2016) and
Behan et al. (2019))
proportion_core_essential_in_essential_I proportion of core essential genes in essential I group
proportion_non_essential_genes_in_essential_II proportion of nonessential genes in essential II group
(nonessential as in Hart and Moffat (2016)); this number should be zero or very close to 0
proportion_non_essential_genes_in_essential_I proportion of nonessential genes in essential I group
score A summary score for the ability of a screen to separate core essential and nonessential genes.
}
\description{
This function extracts sets of
\enumerate{
\item highly essential genes (essentiality level II, the posterior distribution
of the essentiality is shifted less than 1/3 compared to a typical core-essential gene
with with FDR of 0.05, unless a different FDR is specified
\item genes confidently not nonessential with FDR of 0.01, unless a different FDR is specified (essentiality level I)
}
}
\references{
Hart T, Moffat J.
BAGEL: a computational framework for identifying essential genesfrom pooled library screens.
BMC Bioinformatics. 2016;17(1):164.  doi:10.1186/s12859-016-1015-8
Behan FM et al. Prioritization of cancer therapeutic targets using CRISPR–Cas9 screens.
Nature. 2019;568(7753):511–516
}
