% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{extract_gene_categories_FNR}
\alias{extract_gene_categories_FNR}
\title{extract_gene_categories_FNR}
\usage{
extract_gene_categories_FNR(extracted_output, FNR_II = 0.1, FNR_I = 0.1)
}
\arguments{
\item{extracted_output}{output of the function extract_from_output}

\item{FNR_II}{set to 0.05 by default; false negative rate for essentiality of type II}

\item{FNR_I}{set to 0.05 by default}
}
\value{
essential_genes_II  genes of essentiality level II
essential_genes_I genes of essentiality level I
}
\description{
This function is identical to extract_gene_categories except it finds sets of essential genes given a false
negative rather than a false positive rate.
}
\references{
Hart T, Moffat J.
BAGEL: a computational framework for identifying essential genesfrom pooled library screens.
BMC Bioinformatics. 2016;17(1):164.  doi:10.1186/s12859-016-1015-8
Behan FM et al. Prioritization of cancer therapeutic targets using CRISPR–Cas9 screens.
Nature. 2019;568(7753):511–516
}
