% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{fit_bivariate_nv2}
\alias{fit_bivariate_nv2}
\title{fit_bivariate_nv2}
\usage{
fit_bivariate_nv2(
  mean_essential,
  mean_NE,
  sd_essential,
  sd_NE,
  plot_file_name = NA
)
}
\arguments{
\item{mean_essential}{vector of mean log-fold changes for all gRNAs targeting core essential genes,
where the mean is taken across replicates or cell lines and replicates, if there are several cell lines}

\item{mean_NE}{vector of mean log-fold changes for all gRNAs targeting listed nonessential genes}

\item{sd_essential}{standard deviation of log-fold changes for all gRNAs targeting core essential genes}

\item{sd_NE}{standard deviation of log-fold changes for all gRNAs targeting listed nonessential genes}

\item{plot_file}{name (without extension) of the file with the plot of the data and prior distribution}
}
\value{
parameters of bivariate normal prior distribution for mean and standard deviation of gene essentiality
}
\description{
fit_bivariate_nv2
}
\author{
Magdalena Strauss
}
