% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{plot_credible_intervals}
\alias{plot_credible_intervals}
\title{plot_credible_intervals}
\usage{
plot_credible_intervals(extracted_output, name)
}
\arguments{
\item{extracted_output}{output of the function extract_from_output}

\item{name}{name of the screen for naming of output folder and pdf files}
}
\value{
creates pdf files with the following plots:
\enumerate{
\item credible_intervals_as_function_of_mean: credible intervals of essentiality estimates along y-axis,
mean of essentality estimates on x-axis
\item credible_intervals_ordered_by_mean: credible intervals of essentiality estimates along y-axis,
ordered by mean of essentiality estimates (ordered along x-axis)
\item density_length_of_credible_interval: density distribution of the lengths of the credible intervals
\item tail_ESS_credible_interval and Rhat_credible_interval: diagnostic plots to identify any correlation
between convergence measures and the lengths of credible intervals. If there is correlation (which is rare)
this may point to convergence problems of the sampler, and the results may be less reliable.
}
}
\description{
plot_credible_intervals
}
