% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{extract_gene_categories}
\alias{extract_gene_categories}
\title{extract_gene_categories}
\usage{
extract_gene_categories(extracted_output, FDR = 0.05)
}
\arguments{
\item{extracted_output}{output of the function extract_from_output}

\item{FDR}{set to 0.05 by default; maximum false positive rate for all gene sets}
}
\value{
essential_genes_II highly essential genes of essentiality level II (see above)
essential_genes_I genes of essentiality level I (see above), this set should contain almost all of the genes of
essentiality level II. If it does not, this indicates poor screen quality.
proportion_core_essential_in_essential_II which proportion of core essential genes is in the essential genes
II group
(as core essential genes we use the intersection of core essential genes from  Hart and Moffat (2016) and
Behan et al. (2019))
not_essential_genes_II genes that are unessential of type II with a probability with FDR at most 0.05 (unless specified otherwise)
and individual probability at least 0.9
not_essential_genes_I genes that are unessential of type I with a probability with FDR at most 0.05 (unless specified otherwise)
and individual probability at least 0.9
proportion_core_essential_in_essential_I proportion of core essential genes in essential I group
proportion_non_essential_genes_in_essential_II proportion of nonessential genes in essential II group
(nonessential as in Hart and Moffat (2016)); this number should be zero or very close to 0
proportion_non_essential_genes_in_essential_I proportion of nonessential genes in essential I group
score A summary score for the ability of a screen to separate core essential and nonessential genes.
false_negative_rate_essential_II False negative rate for genes essential of type II. This is equal to 0.05 (unless specified otherwise) unless
the individual cutoff of a 90\% probability of being essential type II for each gene was stricter (see above).
false_positive_rate_essential_II False positive rate for genes essential of type II.
false_negative_rate_essential_I False negative rate for genes essential of type I. This is equal to 0.05 (unless specified otherwise) unless
the individual cutoff of a 90\% probability of being essential type I for each gene was stricter (see above).
false_positive_rate_essential_I False positive rate for genes essential of type I.
false_negative_rate_not_essential_II False negative rate for genes unessential of type II. This is equal to 0.05 (unless specified otherwise) unless
the individual cutoff of a 90\% probability of being unessential type II for each gene was stricter (see above).
false_positive_rate_not_essential_II False positive rate for genes unessential of type II.
false_negative_rate_not_essential_I False negative rate for genes unessential of type I. This is equal to 0.05 (unless specified otherwise) unless
the individual cutoff of a 90\% probability of being unessential type I for each gene was stricter (see above).
false_positive_rate_not_essential_I False positive rate for genes unessential of type I.
}
\description{
This function extracts sets of
\enumerate{
\item highly essential genes (essentiality level II, the posterior distribution
of the essentiality is shifted less than 1/3 compared to a typical core-essential gene
with FDR at most 0.05 and individual probability at least 0.9
\item genes confidently not nonessential with FDR at most 0.05 and
individual probability at least 0.9 (essentiality level I)
}
}
\references{
Hart T, Moffat J.
BAGEL: a computational framework for identifying essential genesfrom pooled library screens.
BMC Bioinformatics. 2016;17(1):164.  doi:10.1186/s12859-016-1015-8
Behan FM et al. Prioritization of cancer therapeutic targets using CRISPR–Cas9 screens.
Nature. 2019;568(7753):511–516
}
